fix(travis): avoid 'too long build' error
The Travis build previously failed with 'The job exceeded the maximum time
limit for jobs, and has been terminated'

The Travis build is now split into 2 jobs
  - 1st we build everything but the Studio
  - then build the Studio
Some tricks have been introduced to ensure that the 2 jobs can share the
same cache, in particular, on environment variables declaration.

The build script has been updated to allow to build only the Studio or
skip the Studio build.
As part of this update, the build configuration environment variable names
have changed: they are now prefixed by 'BONITA' instead of 'SCRIPT' for
consistency and better understanding