Remove "generate" profile for Studio build
Generate profile is used to generate classes source code based on some models definition.
A version of this generated source code has been committed in Studio git repository.
So it is not mandatory to regenerate the source code of those classes (except if model definition is edited).
Due to issue https://bonitasoft.atlassian.net/browse/BST-616 this profile need to be removed.
Recommendation if models should be customized is to fork bonita-studio repository, edit the model definitions, generate the source code using Eclipse IDE, commit the new code to the forked repository and update the build script to used the forked repository instead of the official one.

Also update the readme with information about folders to delete to avoid using Tycho (used by Bonita Studio) cache.